build(vue): Rewrite numeral plugin (https://v2.vuejs.org/v2/guide/plugins.html) as useFormatters composable (https://vuejs.org/guide/reusability/composables.html)
